created separate ChessGame class for managing gamestate
