+++ docs/chessgame.js
@@ -1,54 +1,49 @@
-import { ChessMove } from "./chessmove.js";
 import { MoveTracker } from "./movetracker.js";
 import { Piece } from "./piece.js";
-export class WebChessGame {
-    constructor(boardContainer) {
-        this.turncount = 0;
-        this.active = false;
+export class ChessGame {
+    constructor(webgame) {
         this.boardstate = [];
-        this.grid = [];
         this.moveTracker = new MoveTracker();
-        this.boardContainer = boardContainer;
-        this.paintTiles();
-        this.initializeBoardstate();
-        this.fullboardPiecePaint();
+        this.active = false;
+        this.turncount = 0;
+        this.webgame = webgame;
     }
-    checkClickEvent() {
-        const move = this.findClickedCell();
-        if (!move) {
-            return;
-        }
-        if (!this.active) {
-            this.processStartMove(move);
-        }
-        else if (this.active) {
-            this.processEndCell(move);
-            if (this.active) {
-                this.clearHighlights();
-                this.active = false;
-                this.processStartMove(move);
-            }
-        }
+    instantiatePiece(pieceName) {
+        return new Piece(this.webgame, this);
     }
-    setValidMove(i, j) {
-        const tile = this.grid[i][j];
-        if (!tile.classList.contains("validmove")) {
-            this.grid[i][j].classList.add("validmove");
+    lookupPiece(piece) {
+        switch (piece) {
+            case "P":
+                return "pawn";
+            case "R":
+                return "rook";
+            case "N":
+                return "knight";
+            case "B":
+                return "bishop";
+            case "Q":
+                return "queen";
+            case "K":
+                return "king";
         }
+        return "";
     }
-    findClickedCell() {
-        for (let i = 0; i < 8; i++) {
-            for (let j = 0; j < 8; j++) {
-                const tile = this.grid[i][j];
-                if (tile instanceof HTMLElement) {
-                    if (tile.classList.contains("clicked")) {
-                        tile.classList.remove("clicked");
-                        return new ChessMove(i, j);
-                    }
-                }
-            }
-        }
-        return null;
+    activateStart(i, j) {
+        const tile = this.webgame.grid[i][j];
+        this.moveTracker.setStartMove(i, j);
+        tile.classList.add("highlighted");
+        this.active = true;
+        this.populateOptions(i, j);
+    }
+    initializeBoardstate() {
+        this.boardstate.push(["Rb", "Nb", "Bb", "Qb", "Kb", "Bb", "Nb", "Rb"]);
+        this.boardstate.push(["Pb", "Pb", "Pb", "Pb", "Pb", "Pb", "Pb", "Pb"]);
+        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
+        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
+        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
+        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
+        this.boardstate.push(["Pw", "Pw", "Pw", "Pw", "Pw", "Pw", "Pw", "Pw"]);
+        this.boardstate.push(["Rw", "Nw", "Bw", "Qw", "Kw", "Bw", "Nw", "Rw"]);
     }
     getTurnPlayer() {
         if (this.turncount % 2 == 0) {
@@ -63,7 +58,7 @@ export class WebChessGame {
         return true;
     }
     validEnd(i, j) {
-        const tile = this.grid[i][j];
+        const tile = this.webgame.grid[i][j];
         if (tile.classList.contains("validmove")) {
             return true;
         }
@@ -78,15 +73,23 @@ export class WebChessGame {
         if (this.validEnd(move.i, move.j)) {
             this.moveTracker.setEndMove(move.i, move.j);
             this.active = false;
-            this.submitMove();
+            this.webgame.submitMove();
         }
     }
-    activateStart(i, j) {
-        const tile = this.grid[i][j];
-        this.moveTracker.setStartMove(i, j);
-        tile.classList.add("highlighted");
-        this.active = true;
-        this.populateOptions(i, j);
+    legalPosition(i, j, colour) {
+        if (this.validCoordinates(i, j)) {
+            if (this.boardstate[i][j] === ".") {
+                this.webgame.addDot(i, j);
+                return true;
+            }
+            if (this.boardstate[i][j][1] != colour) {
+                this.webgame.addCircle(i, j);
+            }
+        }
+        return false;
+    }
+    validCoordinates(i, j) {
+        return (0 <= i && i < 8 && 0 <= j && j < 8);
     }
     populateOptions(i, j) {
         const pieceChar = this.boardstate[i][j][0];
@@ -107,128 +110,6 @@ export class WebChessGame {
         }
         if (pieceName === "king") {
             piece.kingOptions(i, j, colour);
-        }
-    }
-    instantiatePiece(pieceName) {
-        return new Piece(this);
-    }
-    submitMove() {
-        const startMove = this.moveTracker.getStartMove();
-        const endMove = this.moveTracker.getEndMove();
-        if (!startMove || !endMove) {
-            return;
-        }
-        let a = startMove[0];
-        let b = startMove[1];
-        let x = endMove[0];
-        let y = endMove[1];
-        this.boardstate[x][y] = this.boardstate[a][b];
-        this.boardstate[a][b] = ".";
-        this.paintPosition(x, y);
-        this.paintPosition(a, b);
-        this.clearHighlights();
-        this.turncount += 1;
-    }
-    clearHighlights() {
-        document.querySelectorAll(".highlighted").forEach(el => el.classList.remove("highlighted"));
-        document.querySelectorAll(".validmove").forEach(el => el.classList.remove("validmove"));
-        document.querySelectorAll(".markerdot").forEach(el => el.remove());
-        document.querySelectorAll(".markercircle").forEach(el => el.remove());
-    }
-    legalPosition(i, j, colour) {
-        if (this.validCoordinates(i, j)) {
-            if (this.boardstate[i][j] === ".") {
-                this.addDot(i, j);
-                return true;
-            }
-            if (this.boardstate[i][j][1] != colour) {
-                this.addCircle(i, j);
-            }
-        }
-        return false;
-    }
-    validCoordinates(i, j) {
-        return (0 <= i && i < 8 && 0 <= j && j < 8);
-    }
-    initializeBoardstate() {
-        this.boardstate.push(["Rb", "Nb", "Bb", "Qb", "Kb", "Bb", "Nb", "Rb"]);
-        this.boardstate.push(["Pb", "Pb", "Pb", "Pb", "Pb", "Pb", "Pb", "Pb"]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push(["Pw", "Pw", "Pw", "Pw", "Pw", "Pw", "Pw", "Pw"]);
-        this.boardstate.push(["Rw", "Nw", "Bw", "Qw", "Kw", "Bw", "Nw", "Rw"]);
-    }
-    fullboardPiecePaint() {
-        for (let i = 0; i < 8; i++) {
-            for (let j = 0; j < 8; j++) {
-                this.paintPosition(i, j);
-            }
-        }
-    }
-    addDot(i, j) {
-        const dot = document.createElement("div");
-        dot.classList.add("markerdot");
-        this.setValidMove(i, j);
-        this.grid[i][j].appendChild(dot);
-    }
-    addCircle(i, j) {
-        const circle = document.createElement("div");
-        circle.classList.add("markercircle");
-        this.setValidMove(i, j);
-        this.grid[i][j].appendChild(circle);
-    }
-    lookupPiece(piece) {
-        switch (piece) {
-            case "P":
-                return "pawn";
-            case "R":
-                return "rook";
-            case "N":
-                return "knight";
-            case "B":
-                return "bishop";
-            case "Q":
-                return "queen";
-            case "K":
-                return "king";
-        }
-        return "";
-    }
-    paintPosition(i, j) {
-        const tile = this.grid[i][j];
-        tile.innerHTML = "";
-        if (this.boardstate[i][j] == ".") {
-            return;
-        }
-        const piece = this.boardstate[i][j][0];
-        const pieceName = this.lookupPiece(piece);
-        const colour = this.boardstate[i][j][1];
-        const imgpath = `assets\\${pieceName}_${colour}.png`;
-        const img = document.createElement("img");
-        img.src = imgpath;
-        img.style.margin = "5px 5px";
-        tile.appendChild(img);
-    }
-    paintTiles() {
-        const painting = ["whitebg", "blackbg"];
-        var paint = 0;
-        for (let i = 0; i < 8; i++) {
-            this.grid.push([]);
-            for (let j = 0; j < 8; j++) {
-                const tile = document.createElement("div");
-                tile.classList.add("boardtile");
-                tile.classList.add(painting[paint]);
-                tile.addEventListener("click", () => {
-                    tile.classList.toggle("clicked");
-                    this.checkClickEvent();
-                });
-                this.grid[i].push(tile);
-                this.boardContainer.appendChild(tile);
-                paint = (paint + 1) % 2;
-            }
-            paint = (paint + 1) % 2;
         }
     }
 }
