+++ docs/webchessgame.js
@@ -1,16 +1,16 @@
+import { ChessGame } from "./chessgame.js";
 import { ChessMove } from "./chessmove.js";
 import { MoveTracker } from "./movetracker.js";
-import { Piece } from "./piece.js";
 export class WebChessGame {
     constructor(boardContainer) {
         this.turncount = 0;
         this.active = false;
-        this.boardstate = [];
         this.grid = [];
         this.moveTracker = new MoveTracker();
         this.boardContainer = boardContainer;
+        this.game = new ChessGame(this);
         this.paintTiles();
-        this.initializeBoardstate();
+        this.game.initializeBoardstate();
         this.fullboardPiecePaint();
     }
     checkClickEvent() {
@@ -19,14 +19,14 @@ export class WebChessGame {
             return;
         }
         if (!this.active) {
-            this.processStartMove(move);
+            this.game.processStartMove(move);
         }
         else if (this.active) {
-            this.processEndCell(move);
+            this.game.processEndCell(move);
             if (this.active) {
                 this.clearHighlights();
                 this.active = false;
-                this.processStartMove(move);
+                this.game.processStartMove(move);
             }
         }
     }
@@ -50,68 +50,6 @@ export class WebChessGame {
         }
         return null;
     }
-    getTurnPlayer() {
-        if (this.turncount % 2 == 0) {
-            return "w";
-        }
-        return "b";
-    }
-    validStart(i, j) {
-        if (this.boardstate[i][j] === "." || this.boardstate[i][j][1] != this.getTurnPlayer()) {
-            return false;
-        }
-        return true;
-    }
-    validEnd(i, j) {
-        const tile = this.grid[i][j];
-        if (tile.classList.contains("validmove")) {
-            return true;
-        }
-        return false;
-    }
-    processStartMove(move) {
-        if (this.validStart(move.i, move.j)) {
-            this.activateStart(move.i, move.j);
-        }
-    }
-    processEndCell(move) {
-        if (this.validEnd(move.i, move.j)) {
-            this.moveTracker.setEndMove(move.i, move.j);
-            this.active = false;
-            this.submitMove();
-        }
-    }
-    activateStart(i, j) {
-        const tile = this.grid[i][j];
-        this.moveTracker.setStartMove(i, j);
-        tile.classList.add("highlighted");
-        this.active = true;
-        this.populateOptions(i, j);
-    }
-    populateOptions(i, j) {
-        const pieceChar = this.boardstate[i][j][0];
-        const pieceName = this.lookupPiece(pieceChar);
-        const colour = this.boardstate[i][j][1];
-        const piece = this.instantiatePiece(pieceName);
-        if (pieceName === "pawn") {
-            piece.pawnOptions(i, j, colour);
-        }
-        if (pieceName === "knight") {
-            piece.knightOptions(i, j, colour);
-        }
-        if (pieceName === "rook" || pieceName === "queen") {
-            piece.rookOptions(i, j, colour);
-        }
-        if (pieceName === "bishop" || pieceName === "queen") {
-            piece.bishopOptions(i, j, colour);
-        }
-        if (pieceName === "king") {
-            piece.kingOptions(i, j, colour);
-        }
-    }
-    instantiatePiece(pieceName) {
-        return new Piece(this);
-    }
     submitMove() {
         const startMove = this.moveTracker.getStartMove();
         const endMove = this.moveTracker.getEndMove();
@@ -122,8 +60,8 @@ export class WebChessGame {
         let b = startMove[1];
         let x = endMove[0];
         let y = endMove[1];
-        this.boardstate[x][y] = this.boardstate[a][b];
-        this.boardstate[a][b] = ".";
+        this.game.boardstate[x][y] = this.game.boardstate[a][b];
+        this.game.boardstate[a][b] = ".";
         this.paintPosition(x, y);
         this.paintPosition(a, b);
         this.clearHighlights();
@@ -134,31 +72,6 @@ export class WebChessGame {
         document.querySelectorAll(".validmove").forEach(el => el.classList.remove("validmove"));
         document.querySelectorAll(".markerdot").forEach(el => el.remove());
         document.querySelectorAll(".markercircle").forEach(el => el.remove());
-    }
-    legalPosition(i, j, colour) {
-        if (this.validCoordinates(i, j)) {
-            if (this.boardstate[i][j] === ".") {
-                this.addDot(i, j);
-                return true;
-            }
-            if (this.boardstate[i][j][1] != colour) {
-                this.addCircle(i, j);
-            }
-        }
-        return false;
-    }
-    validCoordinates(i, j) {
-        return (0 <= i && i < 8 && 0 <= j && j < 8);
-    }
-    initializeBoardstate() {
-        this.boardstate.push(["Rb", "Nb", "Bb", "Qb", "Kb", "Bb", "Nb", "Rb"]);
-        this.boardstate.push(["Pb", "Pb", "Pb", "Pb", "Pb", "Pb", "Pb", "Pb"]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push([".", ".", ".", ".", ".", ".", ".", "."]);
-        this.boardstate.push(["Pw", "Pw", "Pw", "Pw", "Pw", "Pw", "Pw", "Pw"]);
-        this.boardstate.push(["Rw", "Nw", "Bw", "Qw", "Kw", "Bw", "Nw", "Rw"]);
     }
     fullboardPiecePaint() {
         for (let i = 0; i < 8; i++) {
@@ -179,32 +92,15 @@ export class WebChessGame {
         this.setValidMove(i, j);
         this.grid[i][j].appendChild(circle);
     }
-    lookupPiece(piece) {
-        switch (piece) {
-            case "P":
-                return "pawn";
-            case "R":
-                return "rook";
-            case "N":
-                return "knight";
-            case "B":
-                return "bishop";
-            case "Q":
-                return "queen";
-            case "K":
-                return "king";
-        }
-        return "";
-    }
     paintPosition(i, j) {
         const tile = this.grid[i][j];
         tile.innerHTML = "";
-        if (this.boardstate[i][j] == ".") {
+        if (this.game.boardstate[i][j] == ".") {
             return;
         }
-        const piece = this.boardstate[i][j][0];
-        const pieceName = this.lookupPiece(piece);
-        const colour = this.boardstate[i][j][1];
+        const piece = this.game.boardstate[i][j][0];
+        const pieceName = this.game.lookupPiece(piece);
+        const colour = this.game.boardstate[i][j][1];
         const imgpath = `assets\\${pieceName}_${colour}.png`;
         const img = document.createElement("img");
         img.src = imgpath;
